Add 'test' workflow and helper scripts
.github/workflows/test.yaml - new test workflow
create_test_dir.sh - script that creates test directory
test - script that runs the test workflow locally with act

The python tests look for a 'test' directory and are executed there. The
bash 'g' script needs to be inside the 'test' directory. This is
inconsistent but not a show stopper.