Allow the user to specify workspace root
In pathological cases '.repo' is not always in workspace root. Therefore
allow the user to specify what file to look for instead of '.repo'.

The user will define the magic file with:
  export WORKSPACE_BOOKMARK_MAGIC_FILE='.wsmagic'

As far as the code is concerned 'path_to' has a new parameter
'magic_file' which is defaulted to '.repo' in order not to break
previously written tests.