Dockerfile: Download bats by hand
The version of bats included in Ubuntu 18.04 is 0.4.0. That is too old
to run the tests. They all simply fail. Locally I have version 1.7.0 so
let's use that in the container too. Since bats is only used for testing
the user doesn't need to download and install bats by hand.