Instead of "run-on" use "container" for testing
When testing locally the helper 'test.sh' replaced "runs-on" with a
custom built container image. This doesn't work on Github, they don't
seem to download the custom container image and instead are stuck with:

  Waiting for a runner to pick up this job...

The correct way of doing things is to specify "runs-on" as any machine
which can run docker that is available in the cloud. That machine then
runs the job's steps in the custom "container".

One curiosity is that by default jobs executed on a runner are using a
bash shell. Jobs executed in the container are using a simple 'sh'.
Since the workspace-bookmark tool uses bash features extensively a bash
compatible shell is required. An example error is:

  | setup.sh: line 0: .: filename argument required
  | .: usage: . filename [arguments]

See https://docs.github.com/en/actions/using-jobs/running-jobs-in-a-container#overview
for more info.