Fix appending "bin" to PATH
The directory 'workspace-bookmark/bin' was appended to PATH by using a
trick which happens to not work with bats. The tests worked because
there was an extra workaround for it in 'src/test_g.bats'.

The issue is that when bats executes tests it will call "load" which
will partially source (See https://github.com/bats-core/bats-core/issues/149)
'setup.sh'. `dirname $0` in 'setup.sh' will get a path to bats'
'libexec/bats-core/bats-exec-test' instead of 'workspace-bookmark/setup.sh'.
This problem does not occur if "BASH_SOURCE[0]" is used instead of "$0".
The downside is that the PATH discovery and amendment mechanizm is now
bash only.