Dockerfile: Install pytest with pip
The version of pytest installed from Ubuntu repositories is quite dated,
it's 3.3.2. In this version the fixture "tmp_path" is unavailable. Since
pytest is required only for testing there is no reason not to upgrade to
a newer version that contains the needed fixture. When installed with
pip version 7.0.1 is installed.