src: Use pytest fixture to create workspace
This has the direct benefit of one now being able to run tests stored in
'test_workspace_bookmark.py' from any directory, which makes setting up
a debug environment much easier.

As part of this change, the fixture which creates a workspace is stored
in 'conftest.py'. This is a magic file which is always auto-imported by
pytest when it runs tests [1].

The previously hardcoded "REPO_DIRECTORY" and other *_DIRECTORY global
variables became test fixtures, the point is that the tests themselves
don't change just yet but are now easier to modify.

[1] https://docs.pytest.org/en/6.2.x/fixture.html#scope-sharing-fixtures-across-classes-modules-packages-or-session