src: Use test fixture for current_location
Up till now the tests hardcoded the starting location from which
"workspace_bookmark.main" is run. To make the tests more resilient to
changes and the testing more thorough the starting locations are
generated randomly but remain in one of the following categories, the
starting ("current") location is either:
* inside a .repo workspace
* outside any workspace

There is one case which the tests don't handle:
* inside magic workspace but outside .repo workspace (due to
  WORKSPACE_BOOKMARKS_MAGIC_FILE being set)
a fixture for this situation will be added when a tests that requires it
is developed.

The test which checks the implementation of support for
WORKSPACE_BOOKMARKS_MAGIC_FILE does a trick... It uses a magic workspace
inside a .repo workspace to validate how bookmarks are transformed into
paths. While technically this does succeed in raising confidence, it
totally misses the point of the original use-case.

The random starting directories may be deeply nested subdirectories not
referenced in WORKSPACE_BOOKMARKS.