src/conftest.py: Create magic workspace fixture
The .repo workspace used as the only workspace for testing is now
created under a magic workspace. It will be now possible to test
previously untested use-cases using the new fixture "magic_workspace".
This includes improving "test_use_magic_file_instead_of_repo" which
succeeds in proving the implementation works as expected but don't
actually test the use-case it was supposed to test.

The fixture "workspace" is renamed to "repo_workspace" to better describe
it's purpose.

The "magic_filename" is a fixture because it will be used for two
purposes:
  1. Provide name of magic directory when construing magic workspace
  2. Populate WORKSPACE_BOOKMARK_MAGIC_FILE
Those are to be done in two separate fixtures but need the same randomly
generated filename.