src: Fix test_use_magic_file_instead_of_repo
The magic directory this test attempts to jump to is now correctly
placed outside the .repo workspace. This matches the expected usage of
WORKSPACE_BOOKMARK_MAGIC_FILE.

In order to achieve this the following fixtures were added:
* magic_directory
* set_magic_file_environment

The assumption made in the fixtures is that magic file is actually a
directory. This is the same assumption the test made previously.