Build a lint docker image
This image will be used to run linters on the source code. It's intended
to be more up to date than the target environments where workspace-bookmark
must run.

The Dockerfile is based on Arch Linux since that is the most up-to-date
Linux distro I'm familiar with.

modified:   .github/workflows/docker-publish.yml
new file:   docker/lint/Dockerfile