src: Rework tests which use "BUILD_DIRECTORY"
The name "BUILD_DIRECTORY" doesn't comply with python's naming scheme
and pylint complains about it [1]. The said fixture is renamed to
"build_directory" to avoid pylint warnings.

Additionally the tests which used "BUILD_DIRECTORY" were slightly
corrected. They now generate a relative path in bookmarks instead of
using absolute paths and relaying on the fact that the "BUILD_DIRECTORY"
is a child of the "POKY_DIRECTORY".

[1] https://github.com/AleksanderBobinski/workspace-bookmark/actions/runs/3915493581/jobs/6693740094