Run black on the python code in CI
Switching between different editors and having code reformatted in
different styles is a pain. Pushing that pain onto others is not a kind
thing to do, hence make sure the style is always the same on the main
branch.

As far as the lint/Dockerfile is concerned arrange the installed
packages in a top-down list in order to make conflict resolution easier,
also git blame will be more helpful now.