Set max-line-length = 88
The default line length used by black for wrapping is 88. Flake8 and
pycodestyle on the other hand warn on anything longer than 79
character. As suggested in [1][2] configure the tools to warn only if
lines exceed 88 characters in order to have slightly more space for
indentations. Now both the formatter and linters are in sync.

Since flake8 reads the local configuration only if it is inside it's cwd
adapt the CI script slightly to start flake8 inside the repo where the
config is defined.

[1] https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html
[2] https://www.youtube.com/watch?v=wf-BqAjZb8M&t=260s